requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
Django==1.8
djangorestframework==3.6.2
requests==2.13.0
wheel==0.24.0
